+++ __init__.py
@@ -2,6 +2,7 @@
 import logging
 from . import aun_lora_info_server as _aun_lora_info_server
 from .AUNAddToPrompt import AUNAddToPrompt
+from .AUNAddToPromptMulti import AUNAddToPromptMulti
 from .AUNAny import AUNAny
 from .AUNBookmark import AUNBookmark
 from .AUNBoolean import AUNBoolean
@@ -85,6 +86,7 @@ WEB_DIRECTORY = "./web"
 
 NODE_CLASS_MAPPINGS = {
     "AUNAddToPrompt": AUNAddToPrompt,
+    "AUNAddToPromptMulti": AUNAddToPromptMulti,
     "AUNAny": AUNAny,
     "AUNBookmark": AUNBookmark,
     "AUNBoolean": AUNBoolean,
@@ -166,6 +168,7 @@ NODE_CLASS_MAPPINGS = {
 
 NODE_DISPLAY_NAME_MAPPINGS = {
     "AUNAddToPrompt": "Add-To-Prompt",
+    "AUNAddToPromptMulti": "Add-To-Prompt (Multi)",
     "AUNAny": "Any",
     "AUNBookmark": "AUN Bookmark",
     "AUNBoolean": "Boolean",